Time: 2 ms (28.75%) | Memory: 40 MB (98.35%) - LeetSync

+++ 2571-find-the-pivot-integer/find-the-pivot-integer.java
@@ -1,8 +1,8 @@
 class Solution {
     public int pivotInteger(int n) {
         for(int x = 1; x<=n; x++){
-            long left_sum = (x*(x+1))/2;
-            long right_sum = (n*(n+1))/2 - (x*(x-1))/2;
+            int left_sum = (x*(x+1))/2;
+            int right_sum = (n*(n+1))/2 - (x*(x-1))/2;
 
             if(left_sum == right_sum) return x;
         }
